requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
bitarray==3.0.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.1
distro==1.9.0
Django==4.2.16
djangorestframework==3.15.2
dnspython==2.6.1
drf-yasg==1.21.8
eventlet==0.37.0
exceptiongroup==1.2.2
flower==2.0.1
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
humanize==4.11.0
idna==3.10
inflection==0.5.1
jiter==0.5.0
joblib==1.4.2
kombu==5.4.2
multidict==6.1.0
mysqlclient==2.2.5
nltk==3.9.1
openai==0.28.0
packaging==24.1
pdfminer.six==20231228
pdfplumber==0.11.4
pdfreader==0.1.15
pillow==10.4.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
PyMuPDF==1.24.11
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.1.0
regex==2024.9.11
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.13.1
